refactor(folders): applied module folder in other files
