+++ src/module/Candidate/components/molecules/AddInformationFields/selectors.ts
@@ -1,4 +1,4 @@
-import { screen } from '~/__stubs__/shared/utils/test-utils'
+import { screen } from '~/shared/utils/test-utils'
 
 export function findByWhereDidYouWorkInput() {
   return screen.findByRole('textbox', { name: 'Onde já trabalhou?' })
