+++ src/module/Candidate/components/molecules/AddInformationFields/stub.d.ts
