+++ src/module/Candidate/components/pages/Main/selectors.ts
@@ -1,4 +1,4 @@
-import { screen } from '~/__stubs__/shared/utils/test-utils'
+import { screen } from '~/shared/utils/test-utils'
 
 export function findByKnowledgeInput() {
   return screen.findByRole('textbox', { name: 'Conhecimentos' })
