+++ src/module/Candidate/store/sagas.ts
@@ -2,9 +2,9 @@ import axios, { AxiosResponse } from 'axios'
 import { call, put, all, takeLatest, select } from 'redux-saga/effects'
 import * as CandidateActions from '~/module/Candidate/store/actions'
 import TYPES from '~/module/Candidate/store/types'
+import { DataProps } from '~/shared/data'
+import Action from '~/shared/lib/typesafe-actions'
 import States from '~/shared/store/rootStates.d'
-import { DataProps } from '~/shared/types/data/index.d'
-import Action from '~/shared/types/lib/typesafe-actions.d'
 
 export function* linkedinDataRequest({ payload: code }: Action<string>) {
   try {
