+++ src/shared/data/LinkedIn.d.ts
