+++ src/shared/data/index.d.ts
