+++ src/shared/utils/test-utils/get-store.ts
@@ -1,7 +1,7 @@
 import { Store } from 'redux'
-import { MyRenderOptions } from '~/__stubs__/shared/utils/test-utils/index.d'
 import { store } from '~/shared/store'
 import rootStates from '~/shared/store/rootStates'
+import { MyRenderOptions } from '~/shared/utils/test-utils/index.d'
 
 const getStore = (options: MyRenderOptions): Store<any, any> => {
   const state = options.initialState || rootStates
